WIP moving logic to monitor
