+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/GoogleServicesDAO.scala
@@ -135,6 +135,10 @@ abstract class GoogleServicesDAO(groupsPrefix: String) extends ErrorReportable {
                                         oldBillingAccount: Option[RawlsBillingAccountName],
                                         force: Boolean = false): Future[ProjectBillingInfo]
 
+  def setBillingAccountName(googleProjectId: GoogleProjectId, billingAccountName: RawlsBillingAccountName): Future[Unit]
+
+  def disableBillingOnGoogleProject(googleProjectId: GoogleProjectId): Future[Unit]
+
   def getBillingInfoForGoogleProject(googleProjectId: GoogleProjectId)(implicit executionContext: ExecutionContext): Future[ProjectBillingInfo]
 
   def getBillingAccountIdForGoogleProject(googleProject: GoogleProject, userInfo: UserInfo)(implicit executionContext: ExecutionContext): Future[Option[String]]
@@ -194,69 +198,6 @@ abstract class GoogleServicesDAO(groupsPrefix: String) extends ErrorReportable {
     *
     */
   def cleanupDMProject(googleProject: GoogleProjectId): Future[Unit]
-
-  def NEW_updateGoogleProjectBillingAccount(googleProjectId: GoogleProjectId,
-                                            newBillingAccount: Option[RawlsBillingAccountName],
-                                            oldBillingAccount: Option[RawlsBillingAccountName],
-                                            force: Boolean = false)(implicit executionContext: ExecutionContext): Future[ProjectBillingInfo] = {
-
-    // Guiding Principle:  Always try to make Terra match what is on Google.  If they fall out of sync with each other,
-    // Google wins.  If we don't know how to reconcile the difference, throw an error.
-    // 1. DO NOT reenable billing if Google says Billing is disabled and Terra does not ALSO think billing is disabled
-    // 2. DO NOT overwrite Billing Account if Terra and Google do not both agree on what the current Billing Account is
-    //    2a. Except if Terra is trying to set the value to match whatever is currently set on Google
-
-    for {
-      projectBillingInfo <- getBillingInfoForGoogleProject(googleProjectId)
-    } yield projectBillingInfo
-  }
-
-  // maybe return a bool or just throw errors
-  def validateBillingAccount(newBillingAccount: Option[RawlsBillingAccountName],
-                             oldBillingAccount: Option[RawlsBillingAccountName],
-                             projectBillingInfo: ProjectBillingInfo,
-                             force: Boolean = false): Boolean = {
-    val currentBillingAccountOnGoogle = if (projectBillingInfo.getBillingAccountName == null || projectBillingInfo.getBillingAccountName.isBlank)
-      None
-    else
-      Option(RawlsBillingAccountName(projectBillingInfo.getBillingAccountName))
-
-    val isBillingEnableOnGoogle = projectBillingInfo.getBillingEnabled
-
-    // Check actual Billing Account value from google against the value that Rawls thinks it should be before the update
-    if (!force
-        && (oldBillingAccount != currentBillingAccountOnGoogle)
-        && (newBillingAccount != currentBillingAccountOnGoogle)) {
-      throw new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.PreconditionFailed,
-        s"Could not update Billing Account on Google Project ID ${projectBillingInfo.getProjectId} to Billing Account ${newBillingAccount} because Billing Account in Rawls ${oldBillingAccount} did not equal current Billing Account in Google ${projectBillingInfo.getBillingAccountName}"))
-    }
-
-    if (!force
-        && (oldBillingAccount.nonEmpty && !isBillingEnableOnGoogle)) {
-      // throw error - Billing is disabled on Google and Terra didn't know about it
-    }
-
-    false
-  }
-
-  // validation should already be done
-  // will determine which call to make to google to update the Billing Account or to disable billing
-  def performUpdate(googleProjectId: GoogleProjectId,
-                    newBillingAccount: Option[RawlsBillingAccountName],
-                    billingAccountOnGoogle: Option[RawlsBillingAccountName]): Future[Unit] = {
-    newBillingAccount match {
-      case Some(billingAccountName) =>
-        if (newBillingAccount != billingAccountOnGoogle) {
-          setBillingAccountName(googleProjectId, billingAccountName)
-        }
-      case None =>
-        disableBillingOnGoogleProject(googleProjectId)
-    }
-  }
-
-  def setBillingAccountName(googleProjectId: GoogleProjectId, billingAccountName: RawlsBillingAccountName): Future[Unit]
-
-  def disableBillingOnGoogleProject(googleProjectId: GoogleProjectId): Future[Unit]
 
   /**
     * Removes the IAM policies from the project's existing policies
