+++ core/src/main/scala/org/broadinstitute/dsde/rawls/monitor/WorkspaceBillingAccountMonitor.scala
@@ -5,10 +5,11 @@ import akka.http.scaladsl.model.StatusCodes
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
 import cats.implicits._
+import com.google.api.services.cloudbilling.model.ProjectBillingInfo
 import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.rawls.RawlsExceptionWithErrorReport
 import org.broadinstitute.dsde.rawls.dataaccess.{GoogleServicesDAO, SlickDataSource}
-import org.broadinstitute.dsde.rawls.model.{GoogleProjectId, RawlsBillingAccountName}
+import org.broadinstitute.dsde.rawls.model.{ErrorReport, GoogleProjectId, RawlsBillingAccountName}
 import org.broadinstitute.dsde.rawls.monitor.WorkspaceBillingAccountMonitor.CheckAll
 
 import scala.concurrent.duration._
@@ -51,6 +52,78 @@ class WorkspaceBillingAccountMonitor(dataSource: SlickDataSource, gcsDAO: Google
           }
       }.unsafeToFuture()
     } yield()
+  }
+
+  def NEW_updateGoogleProjectBillingAccount(googleProjectId: GoogleProjectId,
+                                            newBillingAccount: Option[RawlsBillingAccountName],
+                                            oldBillingAccount: Option[RawlsBillingAccountName],
+                                            force: Boolean = false)(implicit executionContext: ExecutionContext): Future[ProjectBillingInfo] = {
+
+    // Guiding Principle:  Always try to make Terra match what is on Google.  If they fall out of sync with each other,
+    // Google wins.  If we don't know how to reconcile the difference, throw an error.
+    // 1. DO NOT reenable billing if Google says Billing is disabled and Terra does not ALSO think billing is disabled
+    // 2. DO NOT overwrite Billing Account if Terra and Google do not both agree on what the current Billing Account is
+    //    2a. Except if Terra is trying to set the value to match whatever is currently set on Google
+
+    val shouldUpdate, projectBillingInfo = for {
+      projectBillingInfo <- getBillingInfoForGoogleProject(googleProjectId)
+      shouldUpdate = force || validateBillingAccountShouldBeUpdatedInGoogle(newBillingAccount, oldBillingAccount, projectBillingInfo)
+    } yield (shouldUpdate, projectBillingInfo)
+
+
+  }
+
+  def shouldGoogleBeUpdated(oldBillingAccount: Option[RawlsBillingAccountName],
+                            projectBillingInfo: ProjectBillingInfo): Boolean = {
+    val currentBillingAccountOnGoogle = if (projectBillingInfo.getBillingAccountName == null || projectBillingInfo.getBillingAccountName.isBlank)
+      None
+    else
+      Option(RawlsBillingAccountName(projectBillingInfo.getBillingAccountName))
+
+    currentBillingAccountOnGoogle match {
+      case oldBillingAccount => true // google and rawls
+      case _ => false
+    }
+  }
+
+  // maybe return a bool or just throw errors
+  def validateBillingAccountShouldBeUpdatedInGoogle(newBillingAccount: Option[RawlsBillingAccountName],
+                             oldBillingAccount: Option[RawlsBillingAccountName],
+                             projectBillingInfo: ProjectBillingInfo): Boolean = {
+    val currentBillingAccountOnGoogle = if (projectBillingInfo.getBillingAccountName == null || projectBillingInfo.getBillingAccountName.isBlank)
+      None
+    else
+      Option(RawlsBillingAccountName(projectBillingInfo.getBillingAccountName))
+
+    val isBillingEnableOnGoogle = projectBillingInfo.getBillingEnabled
+
+    currentBillingAccountOnGoogle match {
+      case oldBillingAccount => true // google and rawls
+      case newBillingAccount => true // just rawls
+      case _ => // neither
+        throw new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.PreconditionFailed,
+          s"Could not update Billing Account on Google Project ID ${projectBillingInfo.getProjectId} to Billing Account ${newBillingAccount} because Billing Account in Rawls ${oldBillingAccount} did not equal current Billing Account in Google ${projectBillingInfo.getBillingAccountName}"))
+    }
+
+    //    newBillingAccount match {
+    //      case Some(billingAccount) if !isBillingEnableOnGoogle && oldBillingAccount.nonEmpty => // we don't think we need this...?
+    //      case
+    //    }
+  }
+
+  // validation should already be done
+  // will determine which call to make to google to update the Billing Account or to disable billing
+  def performUpdate(googleProjectId: GoogleProjectId,
+                    newBillingAccount: Option[RawlsBillingAccountName],
+                    billingAccountOnGoogle: Option[RawlsBillingAccountName]): Future[Unit] = {
+    newBillingAccount match {
+      case Some(billingAccountName) =>
+        if (newBillingAccount != billingAccountOnGoogle) {
+          setBillingAccountName(googleProjectId, billingAccountName)
+        }
+      case None =>
+        disableBillingOnGoogleProject(googleProjectId)
+    }
   }
 
   private def updateGoogleAndDatabase(googleProjectId: GoogleProjectId,
